Rename patterns to explain, resolving circular import
