+++ src/crested/pl/explain/__init__.py
@@ -1,0 +1,3 @@
+"""Initialize the dist module."""
+
+from ._contribution_scores import contribution_scores
